src/services/matching/requirements.txt
fastapi==0.98.0
python-decouple==3.8
redis==4.5.5
requests==2.31.0
uvicorn==0.22.0
